Invoice Number- Implement updates on invoice
Need to do the following in invoice number

Comment out code for revision invoice

Restriction on duplicate invoice numbers in project

Code level change.
Need to comment out code of revision of invoice number functionality

This functionality needs to be removed

We need to keep the code for enabling it in the future, if needed

The rest functionality should not be impacted.

Just on any updating on the invoice should not update the invoice number to R1 ,R2 etc

This functionality should remain as is When a user creates the first invoice, there will be an additional 01.

Restriction
Need to add restriction to duplicate invoice number.

User should never be able to add manually an invoice number that already exists in the project.

If user has manually added an invoice number and then on another invoice again tries to add same invoice number to any other invoice in the project, it should give error

If an invoice number is autogenerated, the user should not be able to add that number manually to any other invoice.